+++ fourtooncookie-diaryimage-ai-apply-lambda/executer/convert/synopsis_to_scene_convert_executer.py
@@ -1,4 +1,4 @@
-from executer.convert_executer import ConvertExecuter
+from executer.convert.convert_executer import ConvertExecuter
 
 SCENE_PROMPT: str = ""
 SCENE_PROMPT_FILE_PATH = "./prompt/scene_prompt.txt"
