+++ fourtooncookie-diaryimage-ai-apply-lambda/visionrequest/stable_diffusion/executer/scene_as_words_convert_executer.py
@@ -1,4 +1,4 @@
-from executer.convert_executer import ConvertExecuter
+from executer.convert.convert_executer import ConvertExecuter
 
 class SceneAsWordsConvertExecuter(ConvertExecuter):
